backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
werkzeug==3.0.3
python-dotenv==1.0.1
google-genai==1.70.0
gunicorn==22.0.0
